monolitic/requirements.txt
fastapi[standard]>=0.110.0,<0.114.0
uvicorn[standard]>=0.30.0,<0.35.0
scikit-learn
pandas
joblib
yfinance
google-cloud-bigquery>=3.18.0
google-cloud-storage>=2.15.0
fsspec
gcsfs
